+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix processing of chain parameter updates when they occur at the same time.
 - Fix a bug where out-of-band catch-up fails in P6 when processing blocks that have already been
   processed.
+- Fix a bug where certified blocks are not written to disk in some circumstances.
 
 ## 6.0.2